Move getyptes to a domain level function
