+++ src/rest/transactionHandlers.go
@@ -89,39 +89,6 @@ func newTradingPair(tpc managing.TradingPairsManager) func(*fiber.Ctx) error {
 	}
 }
 
-func newAsset(ac managing.AssetsCreator) func(*fiber.Ctx) error {
-	return func(c *fiber.Ctx) error {
-		slog.Info("Creating new asset")
-		payload := struct {
-			Symbol string `form:"symbol"`
-			Name   string `form:"name"`
-			Color  string `form:"color"`
-		}{}
-		if err := c.BodyParser(&payload); err != nil {
-			fmt.Println(err)
-			return err
-		}
-		req := managing.CreateAssetReq{
-			Symbol:      payload.Symbol,
-			Color:       payload.Color,
-			Name:        payload.Name,
-			CountryCode: "-",
-		}
-		resp, err := ac.Create(req)
-		if err != nil {
-			return c.Render("toastErr", fiber.Map{
-				"Title": "Error",
-				"Msg":   err,
-			})
-		}
-		return c.Render("toastOk", fiber.Map{
-			"Title": "Created",
-			"Msg":   resp.Msg,
-		})
-
-	}
-}
-
 func newTransactionImport(tpm managing.TradingPairsManager) func(*fiber.Ctx) error {
 	return func(c *fiber.Ctx) error {
 		slog.Info("Importing transactions")
